chore: reorganize & add logout on frontend
