+++ packages/backend/controllers/classroomController.js
@@ -1,6 +1,4 @@
 const Classroom = require('../models/classroomModel');
-const Student = require('../models/studentModel');
-const Message = require('../models/messageModel');
 const Event = require('../models/eventModel');
 
 const getClassroom = async (req, res) => {
@@ -196,47 +194,10 @@ const getClassroomSubjects = async (req, res) => {
   return res.json(responseData);
 };
 
-const broadcastMessage = async (req, res) => {
-  const { messageHeader, messageBody } = req.body;
-
-  const teacherID = res.locals.user;
-
-  try {
-    const { students } = await Classroom.findOne({ teacher: teacherID });
-
-    if (!students.length) {
-      return res
-        .status(400)
-        .json({ message: 'You have no students in your classroom' });
-    }
-
-    const newMessage = await Message.create({
-      isBroadcast: true,
-      fromTeacher: teacherID,
-      messageHeader,
-      messageBody,
-    });
-
-    students.forEach(async (studentID) => {
-      const student = await Student.findById(studentID);
-
-      student.inbox.push({
-        messageID: newMessage._id,
-        hasBeenRead: false,
-      });
-      await student.save();
-    });
-    return res.status(200).json({ message: 'message sent!' });
-  } catch (error) {
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-};
-
 module.exports = {
   getClassroom,
   updateClassroom,
   deleteClassroom,
-  broadcastMessage,
   getClassroomSubjects,
   getClassroomEvent,
   getClassroomEvents,
