+++ packages/backend/routes/authRoutes.js
@@ -1,7 +1,15 @@
 const router = require('express').Router();
 
-const { logout } = require('../controllers/authController');
+const { authenticateTeacher, authenticateStudent } = require('../middlewares');
 
+const {
+  loginTeacher,
+  loginStudent,
+  logout,
+} = require('../controllers/authController');
+
+router.get('/teacher', authenticateTeacher, loginTeacher);
+router.get('/student', authenticateStudent, loginStudent);
 // "http://localhost:5000/api/v0/auth/logout"
 router.get('/logout', logout);
