+++ packages/backend/routes/index.js
@@ -1,13 +1,13 @@
 const router = require('express').Router();
 
+const authRoutes = require('./authRoutes');
 const teacherRoutes = require('./teacherRoutes');
 const studentRoutes = require('./studentRoutes');
 const classroomRoutes = require('./classroomRoutes');
-const authRoutes = require('./authRoutes');
 
+router.use('/auth', authRoutes);
 router.use('/teacher', teacherRoutes);
 router.use('/student', studentRoutes);
 router.use('/classroom', classroomRoutes);
-router.use('/auth', authRoutes);
 
 module.exports = router;
